Package clean-up. Adding MessageResponse class to handle JSON response.
