+++ src/main/java/com/piekoszek/nowaksharedrent/auth/AccountRepository.java
@@ -1,0 +1,7 @@
+package com.piekoszek.nowaksharedrent.auth;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+interface AccountRepository extends MongoRepository<Account, String> {
+
+}
