+++ src/main/java/com/piekoszek/nowaksharedrent/auth/AuthController.java
@@ -1,0 +1,22 @@
+package com.piekoszek.nowaksharedrent.auth;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+class AuthController {
+
+    private AuthService authService;
+
+    AuthController (AuthService authService) {
+        this.authService = authService;
+    }
+
+    @PostMapping("/auth/signup")
+    ResponseEntity<Object> createAccount(@RequestBody Account account) {
+        if(authService.createAccount(account))
+            return new ResponseEntity<>(new MessageResponse("Account created successfully"), HttpStatus.CREATED);
+        return  new ResponseEntity<>(new MessageResponse("Account with such email already exists"), HttpStatus.BAD_REQUEST);
+    }
+}
