+++ src/main/java/com/piekoszek/nowaksharedrent/auth/AuthService.java
@@ -1,0 +1,5 @@
+package com.piekoszek.nowaksharedrent.auth;
+
+interface AuthService {
+    boolean createAccount(Account account);
+}
