+++ src/main/java/com/piekoszek/nowaksharedrent/auth/AuthServiceImpl.java
@@ -1,0 +1,21 @@
+package com.piekoszek.nowaksharedrent.auth;
+
+import org.springframework.stereotype.Service;
+
+@Service
+class AuthServiceImpl implements AuthService {
+    private AccountRepository accountRepository;
+
+    AuthServiceImpl (AccountRepository accountRepository) {
+        this.accountRepository = accountRepository;
+    }
+
+    @Override
+    public boolean createAccount(Account account) {
+        if(!accountRepository.existsById(account.getEmail())) {
+            accountRepository.save(account);
+            return true;
+        }
+        return false;
+    }
+}
