+++ src/main/java/com/piekoszek/nowaksharedrent/auth/MessageResponse.java
@@ -1,0 +1,13 @@
+package com.piekoszek.nowaksharedrent.auth;
+
+import lombok.Getter;
+
+@Getter
+
+class MessageResponse {
+    String message;
+
+    MessageResponse(String msg) {
+        this.message=msg;
+    }
+}
