--- src/main/java/com/piekoszek/nowaksharedrent/service/AuthService.java
@@ -1,10 +1,0 @@
-package com.piekoszek.nowaksharedrent.service;
-
-import com.piekoszek.nowaksharedrent.model.Account;
-
-import java.util.List;
-
-public interface AuthService {
-    boolean createAccount(Account account);
-    List<Account> getAccount();
-}
